Revert "options: disable vsfilter blur compat by default"
This reverts commit 3d17e19c2c5ca80f916411e7e61126cac8443baa.

This caused files which did not set ScaledBorderAndShadow, and had non-matching
subtitle and video resolutions, to act as if the _display resolution_ was the
video resolution. This meant that those files rendered with a very thin stroke
on displays with resolution higher than the script res.

The performance issues specified in #7435 weren't actually specific to this
case, and can happen with a variety of files.

The option's name is also somewhat misleading: setting the storage size affects
blur, but it also affects stroke, which is far more important for the vast
majority of scripts.